server/nerf-proxy/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
numpy==1.26.4
Pillow==10.2.0

